data_ingestion/pyproject.toml
[project]
name = "data-ingestion"
version = "0.1.0"
description = ""
authors = [
    {name = "angelogagno@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker (>=37.11.0,<38.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
